# openapi.yaml
openapi: 3.0.3
info:
  title: Sprachlern-App API
  description: >
    REST-API für die Sprachlern-App. Deckt Sorgeberechtigte (Guardians) und
    Kinderkonten, Admin-Verwaltung, Lernübungen, Aufgaben, Wörter (mit
    Bild/Audio) sowie den Lernfortschritt ab. Basiert auf dem finalen
    ER-Modell (Version 2) mit Lookup-Tabellen für Sprache, Übungstyp und
    Fortschrittsstatus sowie einer strikten Trennung zwischen
    Guardian-Konten und Admin-Konten (getrennte Tabellen, getrennte
    Authentifizierung).
  version: 2.0.0
servers:
  - url: https://api.sprachlern-app.de/v1
    description: Produktion
  - url: https://staging-api.sprachlern-app.de/v1
    description: Staging
...
paths:
...
  /tasks/{taskId}/words/{wordId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
      - $ref: '#/components/parameters/WordId'
    delete:
      tags: [Tasks]
      summary: Wort-Zuordnung von einer Aufgabe entfernen
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Entfernt
...
  /words:
    get:
      tags: [Words]
      summary: Wörter auflisten
      parameters:
        - name: language_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          description: Volltextsuche auf Wort-Text
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Word'
    # In den "paths:"-Block einfügen, bei den anderen /words/{wordId}/...-Routen:
  /words/{wordId}/tasks:
    get:
      tags: [Words]
      summary: Aufgaben auflisten, die dieses Wort verwenden
      description: >
        Öffentlich lesbar (entspricht dem tatsächlichen Controller-Verhalten;
        vorherige Spec-Version verlangte hier fälschlich adminBearerAuth).
      parameters:
        - name: wordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WordUsage'
        '404':
          description: Wort nicht gefunden
  /words/{wordId}:
    parameters:
      - $ref: '#/components/parameters/WordId'
    get:
      tags: [Words]
      summary: Wort abrufen (inkl. Bild- und Audio-Referenz)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordDetail'
        '404':
          description: Nicht gefunden
    patch:
      tags: [Words]
      summary: Wort aktualisieren
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WordUpdateRequest'
      responses:
        '200':
          description: Aktualisiert
    delete:
      tags: [Words]
      summary: Wort löschen (nur möglich, wenn in keiner Aufgabe verwendet)
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Gelöscht
        '409':
          description: Wort wird noch in mindestens einer Aufgabe verwendet
  /words/{wordId}/audio:
    parameters:
      - $ref: '#/components/parameters/WordId'
    put:
      tags: [Words]
      summary: Audiodatei für ein Wort setzen/ersetzen (1:1)
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudioUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
    delete:
      tags: [Words]
      summary: Audiodatei eines Wortes entfernen
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Gelöscht
  /words/{wordId}/audio/upload:
    parameters:
      - $ref: '#/components/parameters/WordId'
    put:
      tags: [Words]
      summary: Audiodatei für ein Wort hochladen und speichern
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Audio hochgeladen und gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
        '400':
          description: Ungültiger Dateityp oder keine Datei
        '404':
          description: Wort nicht gefunden

  # ============================================================
  # CONCEPTS (sprachübergreifendes Wort: 1 Bild, N Texte/Audios)
  # ============================================================
  /concepts:
    get:
      tags: [Concepts]
      summary: Concepts auflisten
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Concept'
    post:
      tags: [Concepts]
      summary: Neues Concept anlegen (Text für JEDE aktive Sprache erforderlich)
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptCreateRequest'
      responses:
        '201':
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concept'
        '400':
          description: Text für mindestens eine aktive Sprache fehlt, ist doppelt oder unbekannt
  /concepts/{conceptId}:
    parameters:
      - name: conceptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Concepts]
      summary: Einzelnes Concept abrufen
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concept'
        '404':
          description: Nicht gefunden
    delete:
      tags: [Concepts]
      summary: Concept löschen (inkl. aller Sprach-Wörter und des Bilds)
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Gelöscht
  /concepts/{conceptId}/translations:
    parameters:
      - name: conceptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Concepts]
      summary: Text für eine bislang fehlende Sprache nachtragen
      description: >
        Für den Fall, dass eine neue app_language nachträglich eingeführt
        wird und bestehende Concepts dadurch unvollständig werden.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptTranslation'
      responses:
        '201':
          description: Hinzugefügt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concept'
        '409':
          description: Für diese Sprache existiert bereits ein Text
  /concepts/{conceptId}/image:
    parameters:
      - name: conceptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Concepts]
      summary: Bild für ein Concept setzen/ersetzen (gilt für alle Sprachen)
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptImageUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concept'
    delete:
      tags: [Concepts]
      summary: Bild eines Concepts entfernen
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Gelöscht
  /concepts/{conceptId}/image/upload:
    parameters:
      - name: conceptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Concepts]
      summary: Bilddatei für ein Concept hochladen und speichern
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
      responses:
        '200':
          description: Bild hochgeladen und gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concept'
        '400':
          description: Ungültiger Dateityp oder keine Datei
        '404':
          description: Concept nicht gefunden

  # ============================================================
  # FRAGENPOOL (admin-only - jeder Eintrag enthält die Musterlösung)
  # ============================================================
  /question-pool:
    get:
      tags: [QuestionPool]
      summary: Fragenpool auflisten
      security: [{ adminBearerAuth: [] }]
      parameters:
        - name: language_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
    post:
      tags: [QuestionPool]
      summary: Neue Frage (inkl. Musterlösung) anlegen
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
      responses:
        '201':
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /question-pool/{questionId}:
    parameters:
      - name: questionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [QuestionPool]
      summary: Einzelne Frage abrufen
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '404':
          description: Nicht gefunden
    patch:
      tags: [QuestionPool]
      summary: Frage aktualisieren
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdateRequest'
      responses:
        '200':
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
    delete:
      tags: [QuestionPool]
      summary: Frage löschen (nur möglich, wenn in keinem Task verwendet)
      security: [{ adminBearerAuth: [] }]
      responses:
        '204':
          description: Gelöscht
        '409':
          description: Frage wird noch in mindestens einem Task verwendet

  # ============================================================
  # LOOKUPS (Referenzdaten für Dropdowns etc.)
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Guardian-Token, ausgestellt über POST /auth/login. Signiert mit
        JWT_SECRET. Nicht gültig auf Admin-Routen.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Admin-Token, ausgestellt über POST /auth/admin-login. Signiert mit
        einem separaten ADMIN_JWT_SECRET. Nicht gültig auf Guardian-Routen.
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WordId:
      name: wordId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    LookupItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    # ---------- Auth / Guardian ----------
...
    ConceptTranslation:
      type: object
      required: [language_id, text]
      properties:
        language_id:
          type: integer
        text:
          type: string
    ConceptCreateRequest:
      type: object
      required: [translations]
      description: >
        translations MUSS genau einen Eintrag pro aktuell konfigurierter
        app_language enthalten (siehe GET /languages) - sonst 400. Das ist
        die zentrale Regel: ein Concept kann nur vollständig für alle
        Sprachen angelegt werden. Audio wird separat pro Wort über
        PUT /words/{wordId}/audio(/upload) nachgereicht.
      properties:
        translations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ConceptTranslation'
    Concept:
      type: object
      properties:
        concept_id:
          type: string
          format: uuid
        image:
          $ref: '#/components/schemas/Image'
          nullable: true
        words:
          type: array
          items:
            $ref: '#/components/schemas/Word'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConceptImageUpsertRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        description:
          type: string

    # ---------- Fragenpool ----------
    QuestionCreateRequest:
      type: object
      required: [language_id, question, correct_answer]
      properties:
        language_id:
          type: integer
        question:
          type: string
        correct_answer:
          type: string
          description: >
            Musterlösung. Nur über die admin-only /question-pool-Endpoints
            sichtbar, NIE über öffentliche Task-Endpoints.
    QuestionUpdateRequest:
      type: object
      properties:
        language_id:
          type: integer
        question:
          type: string
        correct_answer:
          type: string
    Question:
      type: object
      properties:
        question_id:
          type: string
          format: uuid
        language_id:
          type: integer
        question:
          type: string
        correct_answer:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ---------- Word / Image / Audio ----------
    WordUpdateRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
    Word:
      type: object
      properties:
        word_id:
          type: string
          format: uuid
        concept_id:
          type: string
          format: uuid
        text:
          type: string
        language:
          $ref: '#/components/schemas/LookupItem'
    WordDetail:
      allOf:
        - $ref: '#/components/schemas/Word'
        - type: object
          properties:
            image:
              $ref: '#/components/schemas/Image'
              description: Bild des zugehörigen Concepts (sprachübergreifend geteilt).
            audio:
              $ref: '#/components/schemas/Audio'
    WordUsage:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        question:
          type: string
        exercise_id:
          type: string
          format: uuid
        exercise_title:
          type: string
          nullable: true
    Image:
      type: object
      description: Ein Bild gehört zu einem Concept (sprachübergreifend), nicht zu einem einzelnen Wort.
      properties:
        image_id:
          type: string
          format: uuid
        concept_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        description:
          type: string
        format:
          type: string
          description: Dateiformat laut Cloudinary (z. B. "jpg", "png"). Null bei manuell per URL gesetzten Bildern.
          nullable: true
        bytes:
          type: integer
          description: Dateigröße in Bytes laut Cloudinary. Null bei manuell per URL gesetzten Bildern.
          nullable: true
    AudioUpsertRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        duration_ms:
          type: integer
    Audio:
      type: object
      properties:
        audio_id:
          type: string
          format: uuid
        word_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        duration_ms:
          type: integer
        format:
          type: string
          description: Dateiformat laut Cloudinary (z. B. "mp3", "wav"). Null bei manuell per URL gesetztem Audio.
          nullable: true
        bytes:
          type: integer
          description: Dateigröße in Bytes laut Cloudinary. Null bei manuell per URL gesetztem Audio.
          nullable: true

    # ---------- LearningProgress ----------
...